# api_rede_social/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Protótipo de Rede Social",
        "description": "API protótipo de rede social.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Listar todos os posts com suas tags.",
                "operationId": "8c292c8777126dc6d3efb05793ca8517",
                "responses": {
                    "200": {
                        "description": "Lista de posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID do post",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "usuario_id": {
                                                "description": "ID do usuário",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "description": "Título do post",
                                                "type": "string",
                                                "example": "Meu primeiro post"
                                            },
                                            "content": {
                                                "description": "Conteúdo do post",
                                                "type": "string",
                                                "example": "Conteúdo do meu post."
                                            },
                                            "tags": {
                                                "description": "IDs das tags associadas ao post",
                                                "type": "array",
                                                "items": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Cadastrar post de um usuário com tags.",
                "operationId": "e65ec23954b56b8ece120e19fd94fe87",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "description": "ID do usuário",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "description": "Título do post",
                                        "type": "string",
                                        "example": "Meu primeiro post"
                                    },
                                    "content": {
                                        "description": "Conteúdo do post",
                                        "type": "string",
                                        "example": "Conteúdo do meu post."
                                    },
                                    "tags": {
                                        "description": "IDs das tags associadas ao post, começando do 1",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 6
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do post",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "usuario_id": {
                                            "description": "ID do usuário",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "description": "Título do post",
                                            "type": "string",
                                            "example": "Meu primeiro post"
                                        },
                                        "content": {
                                            "description": "Conteúdo do post",
                                            "type": "string",
                                            "example": "Conteúdo do meu post."
                                        },
                                        "tags": {
                                            "description": "IDs das tags associadas ao post",
                                            "type": "array",
                                            "items": {
                                                "type": "integer",
                                                "example": 2
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou faltando parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar o post."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Exibir post específico por ID.",
                "operationId": "380ebd6ea1942a3eff9d4719a866e457",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do post",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "usuario_id": {
                                            "description": "ID do usuário",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "description": "Título do post",
                                            "type": "string",
                                            "example": "Meu primeiro post"
                                        },
                                        "content": {
                                            "description": "Conteúdo do post",
                                            "type": "string",
                                            "example": "Conteúdo do meu post."
                                        },
                                        "tags": {
                                            "description": "IDs das tags associadas ao post",
                                            "type": "array",
                                            "items": {
                                                "type": "integer",
                                                "example": 2
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Post não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Atualizar post específico.",
                "operationId": "17dddea8ac2c09e99990514c116ef5b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "description": "ID do usuário",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "description": "Título do post",
                                        "type": "string",
                                        "example": "Meu primeiro post"
                                    },
                                    "content": {
                                        "description": "Conteúdo do post",
                                        "type": "string",
                                        "example": "Conteúdo do meu post."
                                    },
                                    "tags": {
                                        "description": "IDs das tags associadas ao post, começando do 1",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 6
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do post",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "usuario_id": {
                                            "description": "ID do usuário",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "description": "Título do post",
                                            "type": "string",
                                            "example": "Meu primeiro post"
                                        },
                                        "content": {
                                            "description": "Conteúdo do post",
                                            "type": "string",
                                            "example": "Conteúdo do meu post."
                                        },
                                        "tags": {
                                            "description": "IDs das tags associadas ao post",
                                            "type": "array",
                                            "items": {
                                                "type": "integer",
                                                "example": 2
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou faltando parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao atualizar o post."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Remover post específico.",
                "operationId": "f43d1f0d1b0f61d306c35cef1ecbf7e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post deletado com sucesso"
                    },
                    "404": {
                        "description": "Post não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Post não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}